@@ -26,7 +26,7 @@
     "crossfilter": "1.x",
     "d3": "3.x",
     "dc": "1.x",
-    "extract-text-webpack-plugin": "^1.0.1",
+    "extract-text-webpack-plugin": "^2.0.0",
     "font-awesome": "4.x",
     "font-awesome-webpack": "0.0.4",
     "immutable": "3.8.1",
@@ -36,23 +36,26 @@
     "less": "^2.7.1",
     "lodash": "3.x",
     "lodash.clonedeep": "^4.5.0",
-    "phylotree": "0.1.4",
+    "phylotree": "0.1.6",
     "react": "15.x.x",
     "react-dom": "^15.4.1"
   },
   "devDependencies": {
     "babel-core": "^6.17.0",
+    "babel-eslint": "^7.2.3",
     "babel-jest": "^19.0.0",
-    "babel-loader": "^6.2.5",
+    "babel-loader": "^7.0.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
     "babel-preset-react-hmre": "^1.1.1",
-    "bootstrap-webpack": "0.0.5",
+    "bootstrap-webpack": "0.0.x",
     "css-loader": "^0.25.0",
-    "eslint": "0.21.x",
+    "eslint": "^4.0.0",
+    "eslint-loader": "^1.8.0",
+    "eslint-plugin-react": "^7.1.0",
     "exports-loader": "^0.6.3",
     "expose-loader": "^0.7.1",
-    "extract-text-webpack-plugin": "^1.0.1",
+    "extract-text-webpack-plugin": "^2.0.0",
     "file-loader": "^0.9.0",
     "gulp": "^3.9.1",
     "gulp-concat": "^2.6.0",
@@ -71,7 +74,8 @@
     "style-loader": "^0.13.1",
     "uglify-js": "2.4.x",
     "url-loader": "^0.5.7",
-    "webpack": "^1.13.2"
+    "webpack": "^3.0.0",
+    "webpack-eslint-plugin": "^1.2.0"
   },
   "scripts": {
     "lint": "eslint dist/hyphyvision.js"
