@@ -1,334 +1,387 @@
-require('dc.css');
+require("dc.css");
 
-var crossfilter = require('crossfilter'),
-		dc = require('dc'),
-	  datamonkey = require('../datamonkey/datamonkey.js');
+var crossfilter = require("crossfilter"),
+  dc = require("dc"),
+  datamonkey = require("../datamonkey/datamonkey.js");
 
 function busted_render_summary(json) {
-
-  var fit_format = d3.format (".2f"),
-      prop_format = d3.format (".2p"),
-      omega_format = d3.format (".3r");
+  var fit_format = d3.format(".2f"),
+    prop_format = d3.format(".2p"),
+    omega_format = d3.format(".3r");
 
   var format_run_time = function(seconds) {
-
     var duration_string = "";
-    seconds = parseFloat (seconds);
-    var split_array = [Math.floor (seconds/(24*3600)) ,Math.floor (seconds/3600) % 24, Math.floor (seconds/60) % 60,seconds % 60],
-        quals = ["d.", "hrs.", "min.", "sec."];
-        
-    split_array.forEach (function (d,i) {
-        if (d) {
-            duration_string += " " + d + " " + quals[i];
-        }
-    });
-    
-    return duration_string;
+    seconds = parseFloat(seconds);
+    var split_array = [
+      Math.floor(seconds / (24 * 3600)),
+      Math.floor(seconds / 3600) % 24,
+      Math.floor(seconds / 60) % 60,
+      seconds % 60
+    ],
+      quals = ["d.", "hrs.", "min.", "sec."];
 
+    split_array.forEach(function(d, i) {
+      if (d) {
+        duration_string += " " + d + " " + quals[i];
+      }
+    });
+
+    return duration_string;
   };
 
   var branch_p_values = {};
-  
+
   var rate_distro_by_branch = {},
-      branch_count = 1,
-      selected_count = 0,
-      tested_count = 0;
-  
+    branch_count = 1,
+    selected_count = 0,
+    tested_count = 0;
+
   var for_branch_table = [];
-  
+
   //var tree_info = render_bs_rel_tree (json, "Unconstrained model");
-  
+
   //var branch_lengths   = tree_info[0],
   //    tested_branches  = {};
-  
+
   for (var p in json["test results"]) {
     branch_p_values[p] = json["test results"]["p"];
     if (branch_p_values[p] <= 0.05) {
-        selected_count++;
+      selected_count++;
     }
   }
-  
-  var fitted_distributions = json["fits"]["Unconstrained model"]["rate distributions"];
-  
+
+  var fitted_distributions =
+    json["fits"]["Unconstrained model"]["rate distributions"];
+
   for (var b in fitted_distributions) {
-     //for_branch_table.push ([b + (tested_branches[b] ? "" : ""),branch_lengths[b],0,0,0]);
-     try {
-          for_branch_table[branch_count][2] = json["test results"][b]["LRT"];
-          for_branch_table[branch_count][3] = json["test results"][b]["p"];
-          for_branch_table[branch_count][4] = json["test results"][b]["uncorrected p"];
-     } catch (e) {
-     }
-     
-     var rateD = fitted_distributions[b];
-     rate_distro_by_branch[b] = rateD; 
-     //for_branch_table[branch_count].push (branch_omegas[b]['distro']);
-     branch_count+=1;
+    //for_branch_table.push ([b + (tested_branches[b] ? "" : ""),branch_lengths[b],0,0,0]);
+    try {
+      for_branch_table[branch_count][2] = json["test results"][b]["LRT"];
+      for_branch_table[branch_count][3] = json["test results"][b]["p"];
+      for_branch_table[branch_count][4] =
+        json["test results"][b]["uncorrected p"];
+    } catch (e) {}
+
+    var rateD = fitted_distributions[b];
+    rate_distro_by_branch[b] = rateD;
+    //for_branch_table[branch_count].push (branch_omegas[b]['distro']);
+    branch_count += 1;
   }
-  
+
   // render summary data
-  var total_tree_length = d3.format("g")(json["fits"]["Unconstrained model"]["tree length"]); 
-  
-  for_branch_table = for_branch_table.sort (function (a,b) {return a[4]-b[4];});
+  for_branch_table = for_branch_table.sort(function(a, b) {
+    return a[4] - b[4];
+  });
 
-  d3.select ('#summary-test-result').text (json['test results']['p'] <= 0.05 ? "evidence" : "no evidence");
-  d3.select ('#summary-test-pvalue').text (d3.format(".3f")(json['test results']['p']));
-  d3.select ('#summary-pmid').text ("PubMed ID " + json['PMID'])
-                             .attr ("href", "http://www.ncbi.nlm.nih.gov/pubmed/" + json['PMID']);
-  d3.select ('#summary-total-runtime').text (format_run_time(json['timers']['overall']));
-  d3.select ('#summary-total-branches').text (branch_count);
-  d3.select ('#summary-tested-branches').text (tested_count);
-  d3.select ('#summary-selected-branches').text (selected_count);
-  
-  has_background = json['background'];
-      
-  var model_rows = [[],[]];
-  
+  d3
+    .select("#summary-test-result")
+    .text(json["test results"]["p"] <= 0.05 ? "evidence" : "no evidence");
+  d3
+    .select("#summary-test-pvalue")
+    .text(d3.format(".3f")(json["test results"]["p"]));
+  d3
+    .select("#summary-pmid")
+    .text("PubMed ID " + json["PMID"])
+    .attr("href", "http://www.ncbi.nlm.nih.gov/pubmed/" + json["PMID"]);
+  d3
+    .select("#summary-total-runtime")
+    .text(format_run_time(json["timers"]["overall"]));
+  d3.select("#summary-total-branches").text(branch_count);
+  d3.select("#summary-tested-branches").text(tested_count);
+  d3.select("#summary-selected-branches").text(selected_count);
+
+  has_background = json["background"];
+
+  var model_rows = [[], []];
+
   if (has_background) {
-      model_rows.push ([]);
+    model_rows.push([]);
   }
-  
-  for (k = 0; k < 2 + has_background; k++)  {
 
-      var access_key,
-          secondary_key,
-          only_distro = 0;
-          
-      if (k === 0) {
+  for (k = 0; k < 2 + has_background; k++) {
+    var access_key,
+      secondary_key,
+      only_distro = 0;
 
-          access_key = 'Unconstrained model';
-          secondary_key = 'FG';
-          model_rows[k].push ('Unconstrained Model');
-          only_distro = 0;
-
+    if (k === 0) {
+      access_key = "Unconstrained model";
+      secondary_key = "FG";
+      model_rows[k].push("Unconstrained Model");
+      only_distro = 0;
+    } else {
+      if (has_background && k == 1) {
+        model_rows[k].push("(background branches)");
+        secondary_key = "BG";
+        only_distro = 1;
       } else {
-
-          if (has_background && k == 1) {
-              model_rows[k].push ('(background branches)');
-              secondary_key = 'BG';
-              only_distro = 1;
-          } else {
-              access_key = 'Constrained model';
-              if (! (access_key in json['fits'])) {
-                break;
-              }
-              model_rows[k].push ('Constrained Model');                    
-              secondary_key = 'FG';
-              only_distro = 0;
-          }
-      }
-      
-      try {
-        model_rows[k].push (only_distro ? '' : fit_format(json['fits'][access_key]['log-likelihood']));
-        model_rows[k].push (only_distro ? '' : json['fits'][access_key]['parameters']);
-        model_rows[k].push (only_distro ? '' : fit_format(json['fits'][access_key]['AIC-c']));
-        model_rows[k].push (only_distro ? '' : format_run_time(json['fits'][access_key]['runtime']));
-        model_rows[k].push (only_distro ? '' : fit_format(json['fits'][access_key]['tree length']));
-
-        for (j = 0; j < 3; j++) {
-         model_rows[k].push (   omega_format(json['fits'][access_key]['rate distributions'][secondary_key][j][0]) + " (" +
-                                prop_format(json['fits'][access_key]['rate distributions'][secondary_key][j][1]) + ")");
+        access_key = "Constrained model";
+        if (!(access_key in json["fits"])) {
+          break;
         }
-      } catch(e) {
-        datamonkey.errorModal(e);
+        model_rows[k].push("Constrained Model");
+        secondary_key = "FG";
+        only_distro = 0;
       }
+    }
+
+    try {
+      model_rows[k].push(
+        only_distro
+          ? ""
+          : fit_format(json["fits"][access_key]["log-likelihood"])
+      );
+      model_rows[k].push(
+        only_distro ? "" : json["fits"][access_key]["parameters"]
+      );
+      model_rows[k].push(
+        only_distro ? "" : fit_format(json["fits"][access_key]["AIC-c"])
+      );
+      model_rows[k].push(
+        only_distro ? "" : format_run_time(json["fits"][access_key]["runtime"])
+      );
+      model_rows[k].push(
+        only_distro ? "" : fit_format(json["fits"][access_key]["tree length"])
+      );
+
+      for (j = 0; j < 3; j++) {
+        model_rows[k].push(
+          omega_format(
+            json["fits"][access_key]["rate distributions"][secondary_key][j][0]
+          ) +
+            " (" +
+            prop_format(
+              json["fits"][access_key]["rate distributions"][secondary_key][
+                j
+              ][1]
+            ) +
+            ")"
+        );
+      }
+    } catch (e) {
+      datamonkey.errorModal(e);
+    }
   }
-                             
-  model_rows = d3.select ('#summary-model-table').selectAll ("tr").data (model_rows);
-  model_rows.enter().append ('tr');
-  model_rows.exit().remove ();
-  model_rows = model_rows.selectAll ("td").data (function (d) {return d;});
-  model_rows.enter().append ("td");
-  model_rows.html (function (d) {return d;});
-  
+
+  model_rows = d3
+    .select("#summary-model-table")
+    .selectAll("tr")
+    .data(model_rows);
+  model_rows.enter().append("tr");
+  model_rows.exit().remove();
+  model_rows = model_rows.selectAll("td").data(function(d) {
+    return d;
+  });
+  model_rows.enter().append("td");
+  model_rows.html(function(d) {
+    return d;
+  });
 }
 
 function busted_render_histogram(c, json) {
-
   var self = this;
 
   // Massage data for use with crossfilter
-  if (d3.keys (json ["evidence ratios"]).length === 0) { // no evidence ratios computed
-    d3.selectAll (c).style ("display", "none");
-    d3.selectAll (".dc-data-table").style ("display", "none");
+  if (d3.keys(json["evidence ratios"]).length === 0) {
+    // no evidence ratios computed
+    d3.selectAll(c).style("display", "none");
+    d3.selectAll(".dc-data-table").style("display", "none");
     //d3.selectAll ('[id^="export"]').style ("display", "none");
-    d3.selectAll ("#er-thresholds").style ("display", "none");
-    d3.selectAll ("#apply-thresholds").style ("display", "none");
+    d3.selectAll("#er-thresholds").style("display", "none");
+    d3.selectAll("#apply-thresholds").style("display", "none");
     return;
   } else {
-    d3.selectAll (c).style ("display", "block");
-    d3.selectAll (".dc-data-table").style ("display", "table");
+    d3.selectAll(c).style("display", "block");
+    d3.selectAll(".dc-data-table").style("display", "table");
     //d3.selectAll ('[id^="export"]').style ("display", "block");
-    d3.selectAll ("#er-thresholds").style ("display", "block");
-    d3.selectAll ("#apply-thresholds").style ("display", "block");
+    d3.selectAll("#er-thresholds").style("display", "block");
+    d3.selectAll("#apply-thresholds").style("display", "block");
   }
 
   var erc = json["evidence ratios"]["constrained"][0];
-  erc = erc.map(function(d) { return Math.log(d)})
-
-  var test_set = json["test set"].split(",");
+  erc = erc.map(function(d) {
+    return Math.log(d);
+  });
   var model_results = [];
 
-  erc.forEach(function(elem, i) { 
+  erc.forEach(function(elem, i) {
     model_results.push({
-      "site_index"          : i + 1,
-      "unconstrained"       : json["profiles"]["unconstrained"][0][i],
-      "constrained"         : json["profiles"]["constrained"][0][i],
-      "optimized_null"      : json["profiles"]["optimized null"][0][i],
-      "er_constrained"      : Math.log(json["evidence ratios"]["constrained"][0][i]),
-      "er_optimized_null"   : Math.log(json["evidence ratios"]["optimized null"][0][i])
-    })
-
+      site_index: i + 1,
+      unconstrained: json["profiles"]["unconstrained"][0][i],
+      constrained: json["profiles"]["constrained"][0][i],
+      optimized_null: json["profiles"]["optimized null"][0][i],
+      er_constrained: Math.log(json["evidence ratios"]["constrained"][0][i]),
+      er_optimized_null: Math.log(
+        json["evidence ratios"]["optimized null"][0][i]
+      )
+    });
   });
 
   var data = crossfilter(model_results);
-  var site_index = data.dimension(function(d) { return d["site_index"]; });
+  var site_index = data.dimension(function(d) {
+    return d["site_index"];
+  });
 
   var sitesByConstrained = site_index.group().reduce(
-    function (p, v) {
+    function(p, v) {
       p.constrained_evidence += +v["er_constrained"];
       p.optimized_null_evidence += +v["er_optimized_null"];
       return p;
     },
-    function (p, v) {
+    function(p, v) {
       p.constrained_evidence -= +v["er_constrained"];
       p.optimized_null_evidence -= +v["er_optimized_null"];
       return p;
     },
-    function () {
-      return { constrained_evidence : 0, optimized_null_evidence : 0 };
+    function() {
+      return { constrained_evidence: 0, optimized_null_evidence: 0 };
     }
   );
 
   var sitesByON = site_index.group().reduce(
-    function (p, v) {
+    function(p, v) {
       p.optimized_null_evidence += +v["er_optimized_null"];
       return p;
     },
-    function (p, v) {
+    function(p, v) {
       p.optimized_null_evidence -= +v["er_optimized_null"];
       return p;
     },
-    function () {
-      return { optimized_null_evidence : 0 };
+    function() {
+      return { optimized_null_evidence: 0 };
     }
   );
-  
-  // Set up new crossfilter dimensions to slice the table by constrained or ON evidence ratio.
-  var er_constrained = data.dimension(function(d) { return d["er_constrained"]; });
-  var er_optimized_null = data.dimension(function(d) { return d["er_optimized_null"]; });
-  self.er_constrained = er_constrained
-  self.er_optimized_null = er_optimized_null
 
+  // Set up new crossfilter dimensions to slice the table by constrained or ON evidence ratio.
+  var er_constrained = data.dimension(function(d) {
+    return d["er_constrained"];
+  });
+  var er_optimized_null = data.dimension(function(d) {
+    return d["er_optimized_null"];
+  });
+  self.er_constrained = er_constrained;
+  self.er_optimized_null = er_optimized_null;
 
   var composite = dc.compositeChart(c);
 
   composite
-      .width($(window).width()*.8)
-      .height(300)
-      .dimension(site_index)
-      .x(d3.scale.linear().domain([1, erc.length]))
-      .yAxisLabel("2 * Ln Evidence Ratio")
-      .xAxisLabel("Site Location")
-      .legend(dc.legend().x($(window).width()*.8 - 150).y(20).itemHeight(13).gap(5))
-      .renderHorizontalGridLines(true)
-      .compose([
-        dc.lineChart(composite)
-          .group(sitesByConstrained, "Constrained")
-          .colors(d3.scale.ordinal().range(['green']))
-          .valueAccessor(function(d) {
-              return 2 * d.value.constrained_evidence;
-          })
-          .keyAccessor(function(d) {
-              return d.key;
-          }), 
-        dc.lineChart(composite)
-          .group(sitesByON, "Optimized Null")
-          .valueAccessor(function(d) {
-              return 2 * d.value.optimized_null_evidence;
-          })
-          .keyAccessor(function(d) {
-              return d.key;
-          })
-          .colors(d3.scale.ordinal().range(['red']))
-      ]);
+    .width($(window).width() * 0.8)
+    .height(300)
+    .dimension(site_index)
+    .x(d3.scale.linear().domain([1, erc.length]))
+    .yAxisLabel("2 * Ln Evidence Ratio")
+    .xAxisLabel("Site Location")
+    .legend(
+      dc.legend().x($(window).width() * 0.8 - 150).y(20).itemHeight(13).gap(5)
+    )
+    .renderHorizontalGridLines(true)
+    .compose([
+      dc
+        .lineChart(composite)
+        .group(sitesByConstrained, "Constrained")
+        .colors(d3.scale.ordinal().range(["green"]))
+        .valueAccessor(function(d) {
+          return 2 * d.value.constrained_evidence;
+        })
+        .keyAccessor(function(d) {
+          return d.key;
+        }),
+      dc
+        .lineChart(composite)
+        .group(sitesByON, "Optimized Null")
+        .valueAccessor(function(d) {
+          return 2 * d.value.optimized_null_evidence;
+        })
+        .keyAccessor(function(d) {
+          return d.key;
+        })
+        .colors(d3.scale.ordinal().range(["red"]))
+    ]);
 
   composite.xAxis().ticks(50);
 
   var numberFormat = d3.format(".4f");
 
   // Render the table
-  dc.dataTable(".dc-data-table")
-      .dimension(site_index)
-      // data table does not use crossfilter group but rather a closure
-      // as a grouping function
-      .group(function (d) {
-        return site_index.bottom(1)[0].site_index + " - " + site_index.top(1)[0].site_index;
-      })
-      .size(site_index.groupAll().reduceCount().value()) // (optional) max number of records to be shown, :default = 25
-      // dynamic columns creation using an array of closures
-      .columns([
-          function (d) {
-              return d.site_index;
-          },
-          function (d) {
-              return numberFormat(d["unconstrained"]);
-          },
-          function (d) {
-              return numberFormat(d["constrained"]);
-          },
-          function (d) {
-              return numberFormat(d["optimized_null"]);
-          },
-          function (d) {
-              return numberFormat(d["er_constrained"]);
-          },
-          function (d) {
-              return numberFormat(d["er_optimized_null"]);
-          },
+  dc
+    .dataTable(".dc-data-table")
+    .dimension(site_index)
+    // data table does not use crossfilter group but rather a closure
+    // as a grouping function
+    .group(function(d) {
+      return (
+        site_index.bottom(1)[0].site_index +
+        " - " +
+        site_index.top(1)[0].site_index
+      );
+    })
+    .size(site_index.groupAll().reduceCount().value()) // (optional) max number of records to be shown, :default = 25
+    // dynamic columns creation using an array of closures
+    .columns([
+      function(d) {
+        return d.site_index;
+      },
+      function(d) {
+        return numberFormat(d["unconstrained"]);
+      },
+      function(d) {
+        return numberFormat(d["constrained"]);
+      },
+      function(d) {
+        return numberFormat(d["optimized_null"]);
+      },
+      function(d) {
+        return numberFormat(d["er_constrained"]);
+      },
+      function(d) {
+        return numberFormat(d["er_optimized_null"]);
+      }
+    ])
+    // (optional) sort using the given field, :default = function(d){return d;}
+    .sortBy(function(d) {
+      return d.site_index;
+    })
+    // (optional) sort order, :default ascending
+    .order(d3.ascending)
+    // (optional) custom renderlet to post-process chart using D3
+    .renderlet(function(table) {
+      table.selectAll(".dc-table-group").classed("info", true);
+    });
 
-      ])
+  $("#export-csv").on("click", function(e) {
+    datamonkey.export_csv_button(site_index.top(Infinity));
+  });
 
-      // (optional) sort using the given field, :default = function(d){return d;}
-      .sortBy(function (d) {
-          return d.site_index;
-      })
-
-      // (optional) sort order, :default ascending
-      .order(d3.ascending)
-
-      // (optional) custom renderlet to post-process chart using D3
-      .renderlet(function (table) {
-          table.selectAll(".dc-table-group").classed("info", true);
-      });
-
-  $("#export-csv").on('click', function(e) { datamonkey.export_csv_button(site_index.top(Infinity)); } );
-
-  $("#export-chart-svg").on('click', function(e) { 
+  $("#export-chart-svg").on("click", function(e) {
     // class manipulation for the image to display correctly
     $("#chart-id").find("svg")[0].setAttribute("class", "dc-chart");
-    datamonkey.save_image("svg", "#chart-id"); 
+    datamonkey.save_image("svg", "#chart-id");
     $("#chart-id").find("svg")[0].setAttribute("class", "");
   });
 
-  $("#export-chart-png").on('click', function(e) { 
+  $("#export-chart-png").on("click", function(e) {
     // class manipulation for the image to display correctly
     $("#chart-id").find("svg")[0].setAttribute("class", "dc-chart");
-    datamonkey.save_image("png", "#chart-id"); 
+    datamonkey.save_image("png", "#chart-id");
     $("#chart-id").find("svg")[0].setAttribute("class", "");
   });
-  $("#apply-thresholds").on('click', function(e) { 
-    var erConstrainedThreshold = document.getElementById("er-constrained-threshold").value;
-    var erOptimizedNullThreshold = document.getElementById("er-optimized-null-threshold").value;
-    self.er_constrained.filter(function(d) { return d >= erConstrainedThreshold; });
-    self.er_optimized_null.filter(function(d) { return d >= erOptimizedNullThreshold; });
+  $("#apply-thresholds").on("click", function(e) {
+    var erConstrainedThreshold = document.getElementById(
+      "er-constrained-threshold"
+    ).value;
+    var erOptimizedNullThreshold = document.getElementById(
+      "er-optimized-null-threshold"
+    ).value;
+    self.er_constrained.filter(function(d) {
+      return d >= erConstrainedThreshold;
+    });
+    self.er_optimized_null.filter(function(d) {
+      return d >= erOptimizedNullThreshold;
+    });
     dc.renderAll();
   });
 
-
   dc.renderAll();
-
 }
 
 module.exports.render_summary = busted_render_summary;
 module.exports.render_histogram = busted_render_histogram;
-
