@@ -1,5 +1,4 @@
 function getStyles(doc) {
-
   function processStyleSheet(ss) {
     if (ss.cssRules) {
       for (var i = 0; i < ss.cssRules.length; i++) {
@@ -20,7 +19,7 @@ function getStyles(doc) {
   }
 
   var styles = "",
-      styleSheets = doc.styleSheets;
+    styleSheets = doc.styleSheets;
 
   if (styleSheets) {
     for (var i = 0; i < styleSheets.length; i++) {
@@ -29,27 +28,27 @@ function getStyles(doc) {
   }
 
   return styles;
-
 }
 
 function exportCSVButton(data) {
-
   data = d3.csv.format(data);
   if (data !== null) {
-    var pom = document.createElement('a');
-    pom.setAttribute('href', 'data:text/csv;charset=utf-8,' + encodeURIComponent(data));
-    pom.setAttribute('download', 'export.csv');
-    pom.className = 'btn btn-default btn-sm';
-    pom.innerHTML = '<span class="glyphicon glyphicon-floppy-save"></span> Download CSV';
+    var pom = document.createElement("a");
+    pom.setAttribute(
+      "href",
+      "data:text/csv;charset=utf-8," + encodeURIComponent(data)
+    );
+    pom.setAttribute("download", "export.csv");
+    pom.className = "btn btn-default btn-sm";
+    pom.innerHTML =
+      '<span class="glyphicon glyphicon-floppy-save"></span> Download CSV';
     $("body").append(pom);
     pom.click();
     pom.remove();
   }
-
 }
 
 function convertSVGtoPNG(image_string) {
-
   var image = document.getElementById("chart-image");
   image.src = image_string;
 
@@ -59,27 +58,25 @@ function convertSVGtoPNG(image_string) {
     canvas.height = image.height;
     var context = canvas.getContext("2d");
     context.fillStyle = "#FFFFFF";
-    context.fillRect(0,0,image.width,image.height);
+    context.fillRect(0, 0, image.width, image.height);
     context.drawImage(image, 0, 0);
     var img = canvas.toDataURL("image/png");
 
-    var pom = document.createElement('a');
-    pom.setAttribute('download', 'image.png');
-    pom.href = canvas.toDataURL("image/png");     
+    var pom = document.createElement("a");
+    pom.setAttribute("download", "image.png");
+    pom.href = canvas.toDataURL("image/png");
     $("body").append(pom);
     pom.click();
     pom.remove();
-  }
-
+  };
 }
 
 function saveImage(type, container) {
-
   var prefix = {
     xmlns: "http://www.w3.org/2000/xmlns/",
     xlink: "http://www.w3.org/1999/xlink",
     svg: "http://www.w3.org/2000/svg"
-  }
+  };
 
   var svg = $(container).find("svg")[0];
   var styles = getStyles(window.document);
@@ -87,12 +84,11 @@ function saveImage(type, container) {
   svg.setAttribute("version", "1.1");
 
   var defsEl = document.createElement("defs");
-  svg.insertBefore(defsEl, svg.firstChild); 
+  svg.insertBefore(defsEl, svg.firstChild);
 
   var styleEl = document.createElement("style");
   defsEl.appendChild(styleEl);
   styleEl.setAttribute("type", "text/css");
-
 
   // removing attributes so they aren't doubled up
   svg.removeAttribute("xmlns");
@@ -107,21 +103,24 @@ function saveImage(type, container) {
     svg.setAttributeNS(prefix.xmlns, "xmlns:xlink", prefix.xlink);
   }
 
-  var source = (new XMLSerializer()).serializeToString(svg).replace('</style>', '<![CDATA[' + styles + ']]></style>');
+  var source = new XMLSerializer()
+    .serializeToString(svg)
+    .replace("</style>", "<![CDATA[" + styles + "]]></style>");
   var rect = svg.getBoundingClientRect();
-  var doctype = '<?xml version="1.0" standalone="no"?><!DOCTYPE svg PUBLIC "-//W3C//DTD SVG 1.1//EN" "http://www.w3.org/Graphics/SVG/1.1/DTD/svg11.dtd">';
+  var doctype =
+    '<?xml version="1.0" standalone="no"?><!DOCTYPE svg PUBLIC "-//W3C//DTD SVG 1.1//EN" "http://www.w3.org/Graphics/SVG/1.1/DTD/svg11.dtd">';
   var to_download = [doctype + source];
-  var image_string = 'data:image/svg+xml;base66,' + encodeURIComponent(to_download);
+  var image_string =
+    "data:image/svg+xml;base66," + encodeURIComponent(to_download);
 
-  if(type == "png") {
+  if (type == "png") {
     convertSVGtoPNG(image_string);
   } else {
-    var pom = document.createElement('a');
-    pom.setAttribute('download', 'image.svg');
-    pom.setAttribute('href', image_string);
+    var pom = document.createElement("a");
+    pom.setAttribute("download", "image.svg");
+    pom.setAttribute("href", image_string);
     $("body").append(pom);
     pom.click();
     pom.remove();
   }
-
 }
