@@ -1,21 +1,20 @@
-require('datamonkey.js');
+require("datamonkey.js");
 
 function datamonkey_alignment_viewer(options) {
-
   var self = this;
 
- function initialize(options) {
+  function initialize(options) {
     if (!options) options = {};
 
     //-- Public
-    self.container = $('div.seq-viewer');
+    self.container = $("div.seq-viewer");
     self.dataset = [];
     self.cell_size = 30;
     self.zoom = { low: 0.0, high: 1.0 };
 
     self.dataset = [
       "TREESPARROW_HENAN_1_2004",
-      "HUMAN_VIETNAM_CL105_2005", 
+      "HUMAN_VIETNAM_CL105_2005",
       "TREESPARROW_HENAN_4_2004",
       "CHICKEN_HEBEI_326_2005",
       "CHICKEN_HONGKONG_915_97",
@@ -38,125 +37,129 @@ function datamonkey_alignment_viewer(options) {
     ];
 
     //-- Private
-    //self.xScale; 
+    //self.xScale;
     //self.yScale;
     //self.mean = 0;
     //self.tooltip;
     //self.infotip;
 
     draw();
-
   }
-
 
   function draw() {
-
     var width = $(window).width();
-    var height = $(window).height() - Math.floor($('.navbar').height()*1.1);
+    var height = $(window).height() - Math.floor($(".navbar").height() * 1.1);
     var offset = 15;
-  
+
     //initial draw
-    self.paper = d3.selectAll( self.container.selector )
-      .append('svg')
-        .attr("width", width)
-        .attr("height", height);
+    self.paper = d3
+      .selectAll(self.container.selector)
+      .append("svg")
+      .attr("width", width)
+      .attr("height", height);
 
-    // Y Legend 
-    var yAxisAttrs = { 'class'  : 'yAxis',
-                       'height' : height,
-                       'width'  : width
-                     };
+    // Y Legend
+    var yAxisAttrs = {
+      class: "yAxis",
+      height: height,
+      width: width
+    };
 
-    self.yAxis = self.paper
-      .append('g')
-        .attr(yAxisAttrs);
+    self.yAxis = self.paper.append("g").attr(yAxisAttrs);
 
-
-    self.yAxisLabels = self.yAxis
-      .append('g')
-        .attr('class', 'sequence-labels');
+    self.yAxisLabels = self.yAxis.append("g").attr("class", "sequence-labels");
 
     var legenedHLabel = self.yAxisLabels
-        .selectAll('sequence-label')
-          .data(self.dataset);
+      .selectAll("sequence-label")
+      .data(self.dataset);
 
-    var legenedHLabelAttrs = { 'class' : 'sequence-label',
-                               'y' : function (d, i) { return i * self.cell_size + offset; },
-                               'fill' : '#333',
-                               'dominant-baseline' : 'central',
-                               'text-anchor' : 'right',
-                               'font-family' : '"Helvetica Neue", Helvetica, Arial, "Lucida Grande", sans-serif', };
+    var legenedHLabelAttrs = {
+      class: "sequence-label",
+      y: function(d, i) {
+        return i * self.cell_size + offset;
+      },
+      fill: "#333",
+      "dominant-baseline": "central",
+      "text-anchor": "right",
+      "font-family":
+        '"Helvetica Neue", Helvetica, Arial, "Lucida Grande", sans-serif'
+    };
 
-    legenedHLabel.enter()
-      .append('text')
-        .attr(legenedHLabelAttrs)
-        .text(function(d) { return d; });
+    legenedHLabel
+      .enter()
+      .append("text")
+      .attr(legenedHLabelAttrs)
+      .text(function(d) {
+        return d;
+      });
 
-    legenedHLabel.exit()
-      .remove();
+    legenedHLabel.exit().remove();
 
-    self.xScale = d3.scale.linear()
-        .domain([
-                (function() {
-                  if(d3.min(values) < 0) {
-                    return (d3.min(values));
-                  } else {
-                    return (d3.max(values));
-                  }
-                })(),
+    self.xScale = d3.scale
+      .linear()
+      .domain([
+        (function() {
+          if (d3.min(values) < 0) {
+            return d3.min(values);
+          } else {
+            return d3.max(values);
+          }
+        })(),
 
-                 (d3.max(values) + maxMean)*self.zoom.high
-                 ])
-        .range([0, width]);
+        (d3.max(values) + maxMean) * self.zoom.high
+      ])
+      .range([0, width]);
 
+    // X Legend
+    self.xAxis = d3.svg.axis().scale(self.xScale).orient("top");
 
-    // X Legend 
-    self.xAxis = d3.svg.axis()
-        .scale(self.xScale)
-        .orient('top')
+    var xAxisHeaderAttrs = {
+      class: "xAxis header",
+      height: 20,
+      width: 2000,
+      transform: "translate(" + width + ", 0)"
+    };
 
-    var xAxisHeaderAttrs = { 'class'     : 'xAxis header',
-                             'height'    : 20,
-                             'width'     : 2000,
-                             'transform' : 'translate('+ (width) +', 0)', 
-                           }
+    self.xAxisHeader = self.paper.append("g").attr(xAxisHeaderAttrs);
 
-    self.xAxisHeader = self.paper
-      .append('g')
-        .attr(xAxisHeaderAttrs);
+    self.xAxisHeader.style("fill", "none").call(self.xAxis);
 
-    self.xAxisHeader
-      .style('fill', 'none')
-      .call(self.xAxis);
-
-    self.yScale = d3.scale.ordinal()
-      .domain( d3.range(self.dataset.length) )
+    self.yScale = d3.scale
+      .ordinal()
+      .domain(d3.range(self.dataset.length))
       .rangeRoundBands([0, height], 0.25);
 
-    var xGridAttrs = { 'transform' : 'translate(0,'+ self.yScale(self.dataset.length) +')',
-                       'fill' : '#ccc',
-                       'font-size' : '12px',
-                       'font-family' : '"Helvetica Neue", Helvetica, Arial, "Lucida Grande", sans-serif', };
+    var xGridAttrs = {
+      transform: "translate(0," + self.yScale(self.dataset.length) + ")",
+      fill: "#ccc",
+      "font-size": "12px",
+      "font-family":
+        '"Helvetica Neue", Helvetica, Arial, "Lucida Grande", sans-serif'
+    };
 
     self.xAxisHeader
-      .append('g')
-        .attr('class', 'guide_lines_x')
-        .call( self.xAxis.tickFormat('').tickSize(-height,0,0));
+      .append("g")
+      .attr("class", "guide_lines_x")
+      .call(self.xAxis.tickFormat("").tickSize(-height, 0, 0));
 
-    var xAxisHeaderStyles = { 'fill' : '#000',
-                              'font-size' : 12,
-                              'font-family' : '"Helvetica Neue", Helvetica, Arial, "Lucida Grande", sans-serif'};
+    var xAxisHeaderStyles = {
+      fill: "#000",
+      "font-size": 12,
+      "font-family":
+        '"Helvetica Neue", Helvetica, Arial, "Lucida Grande", sans-serif'
+    };
 
-    self.xAxisHeader.selectAll('g.xAxis.header g text').style(xAxisHeaderStyles);
-    var xAxisTickLineStyles = { 'stroke' : '#CCC', };
-    self.xAxisHeader.selectAll('g.guide_lines_x g line').style(xAxisTickLineStyles);
-
+    self.xAxisHeader
+      .selectAll("g.xAxis.header g text")
+      .style(xAxisHeaderStyles);
+    var xAxisTickLineStyles = { stroke: "#CCC" };
+    self.xAxisHeader
+      .selectAll("g.guide_lines_x g line")
+      .style(xAxisTickLineStyles);
   }
-
 
   initialize(options);
   return this;
-
 }
 
 datamonkey.alignment_viewer = datamonkey_alignment_viewer;
