@@ -2,57 +2,73 @@ require("jquery");
 require("datamonkey.js");
 
 function datamonkey_cancel(self) {
-    $(self).show().next().remove();
+  $(self).show().next().remove();
 }
 
 function datamonkey_change(value, self) {
-    $(self).text (value)
-        .show()
-        .next().remove();
+  $(self).text(value).show().next().remove();
 }
 
-function datamonkey_check_valid_value (value, value_list, previous) {
-    if (value.length) {
-        if (value == previous) {
-            return true;
-        }
-        if (value_list) {
-            return !(value in value_list);
-        }
-        return true;
+function datamonkey_check_valid_value(value, value_list, previous) {
+  if (value.length) {
+    if (value == previous) {
+      return true;
     }
-    return false;
+    if (value_list) {
+      return !(value in value_list);
+    }
+    return true;
+  }
+  return false;
 }
 
 function datamonkey_editable(self, value_list, edit_group) {
-    $(self).hide();
-    
-    if (edit_group) {
-        edit_group.filter (function (d) {return d[1];}).forEach (function (d) {datamonkey_cancel (d[0]);});
-        edit_group.forEach (function (d) {d[1] = d[0] === self;});
-    }
+  $(self).hide();
 
-    var div            = d3.select ($(self).parent()[0]).append ("div").classed ("input-group", true);
-        text_field     = div.append ("input").style ("margin-right","1em"),
-        button_ok      = div.append ("button").classed ("btn btn-primary btn-xs", true),
-        button_cancel  = div.append ("button").classed ("btn btn-primary btn-xs", true),
-        current_value  = $(self).text();
+  if (edit_group) {
+    edit_group
+      .filter(function(d) {
+        return d[1];
+      })
+      .forEach(function(d) {
+        datamonkey_cancel(d[0]);
+      });
+    edit_group.forEach(function(d) {
+      d[1] = d[0] === self;
+    });
+  }
 
-    button_ok.append ("i").classed ("glyphicon glyphicon-ok", true);
-    button_cancel.append ("i").classed ("glyphicon glyphicon-remove", true);
+  var div = d3
+    .select($(self).parent()[0])
+    .append("div")
+    .classed("input-group", true);
+  (text_field = div
+    .append("input")
+    .style("margin-right", "1em")), (button_ok = div
+    .append("button")
+    .classed("btn btn-primary btn-xs", true)), (button_cancel = div
+    .append("button")
+    .classed("btn btn-primary btn-xs", true)), (current_value = $(self).text());
 
+  button_ok.append("i").classed("glyphicon glyphicon-ok", true);
+  button_cancel.append("i").classed("glyphicon glyphicon-remove", true);
 
-    $(text_field[0]).val(current_value).on ("input propertychange", function (event) {
-        button_ok.property ("disabled", !datamonkey_check_valid_value ($(this).val(), value_list,current_value));
+  $(text_field[0])
+    .val(current_value)
+    .on("input propertychange", function(event) {
+      button_ok.property(
+        "disabled",
+        !datamonkey_check_valid_value($(this).val(), value_list, current_value)
+      );
     });
 
-    button_ok.on ("click", function (event, datum) {
-        datamonkey_change ($(text_field[0]).val(), self);
-    });
+  button_ok.on("click", function(event, datum) {
+    datamonkey_change($(text_field[0]).val(), self);
+  });
 
-    button_cancel.on ("click", function (event, datum) {
-        datamonkey_cancel (self);
-    });
+  button_cancel.on("click", function(event, datum) {
+    datamonkey_cancel(self);
+  });
 }
 
 datamonkey.editable = datamonkey_editable;
