@@ -1,58 +1,59 @@
 require("./datamonkey.js");
 
 function set_tree_handlers(tree_object) {
-    $("[data-direction]").on("click", function(e) {
-        var which_function = $(this).data("direction") == 'vertical' ? tree_object.spacing_x : tree_object.spacing_y;
-        which_function(which_function() + (+$(this).data("amount"))).update();
-    });
+  $("[data-direction]").on("click", function(e) {
+    var which_function = $(this).data("direction") == "vertical"
+      ? tree_object.spacing_x
+      : tree_object.spacing_y;
+    which_function(which_function() + +$(this).data("amount")).update();
+  });
 
-    $(".phylotree-layout-mode").on("change", function(e) {
-        if ($(this).is(':checked')) {
-            if (tree_object.radial() != ($(this).data("mode") == "radial")) {
-                tree_object.radial(!tree_object.radial()).placenodes().update();
-            }
-        }
-    });
-
-    $(".phylotree-align-toggler").on("change", function(e) {
-        if ($(this).is(':checked')) {
-            if (tree_object.align_tips($(this).data("align") == "right")) {
-                tree_object.placenodes().update();
-            }
-        }
-    });
-
-    $("#sort_original").on("click", function(e) {
-        tree_object.resort_children(function(a, b) {
-            return a["original_child_order"] - b["original_child_order"];
-        });
-
-        e.preventDefault();
-
-    });
-
-    $("#sort_ascending").on("click", function(e) {
-        sort_nodes(true);
-        e.preventDefault();
-    });
-
-    $("#sort_descending").on("click", function(e) {
-        sort_nodes(false);
-        e.preventDefault();
-    });
-
-    function sort_nodes(asc) {
-        tree_object.traverse_and_compute(function(n) {
-            var d = 1;
-            if (n.children && n.children.length) {
-                d += d3.max(n.children, function(d) {
-                    return d["count_depth"];
-                });
-            }
-            n["count_depth"] = d;
-        });
-        tree_object.resort_children(function(a, b) {
-            return (a["count_depth"] - b["count_depth"]) * (asc ? 1 : -1);
-        });
+  $(".phylotree-layout-mode").on("change", function(e) {
+    if ($(this).is(":checked")) {
+      if (tree_object.radial() != ($(this).data("mode") == "radial")) {
+        tree_object.radial(!tree_object.radial()).placenodes().update();
+      }
     }
+  });
+
+  $(".phylotree-align-toggler").on("change", function(e) {
+    if ($(this).is(":checked")) {
+      if (tree_object.align_tips($(this).data("align") == "right")) {
+        tree_object.placenodes().update();
+      }
+    }
+  });
+
+  $("#sort_original").on("click", function(e) {
+    tree_object.resort_children(function(a, b) {
+      return a["original_child_order"] - b["original_child_order"];
+    });
+
+    e.preventDefault();
+  });
+
+  $("#sort_ascending").on("click", function(e) {
+    sort_nodes(true);
+    e.preventDefault();
+  });
+
+  $("#sort_descending").on("click", function(e) {
+    sort_nodes(false);
+    e.preventDefault();
+  });
+
+  function sort_nodes(asc) {
+    tree_object.traverse_and_compute(function(n) {
+      var d = 1;
+      if (n.children && n.children.length) {
+        d += d3.max(n.children, function(d) {
+          return d["count_depth"];
+        });
+      }
+      n["count_depth"] = d;
+    });
+    tree_object.resort_children(function(a, b) {
+      return (a["count_depth"] - b["count_depth"]) * (asc ? 1 : -1);
+    });
+  }
 }
