@@ -6,12 +6,13 @@ require('./fade/FADE.css');
 require('bootstrap');
 require('./datamonkey/datamonkey.js');
 
-var absrel = require('./jsx/absrel.jsx');
-var busted = require('./jsx/busted.jsx');
-var fade = require('./fade/FADE.js');
-var fade_summary = require('./jsx/fade_summary.jsx');
-var relax = require('./jsx/relax.jsx');
-var slac = require('./jsx/slac.jsx');
+var absrel = require('./jsx/absrel.jsx'),
+    busted = require('./jsx/busted.jsx'),
+    fade = require('./fade/FADE.js'),
+    fade_summary = require('./jsx/fade_summary.jsx'),
+    fel = require('./jsx/fel.jsx'),
+    relax = require('./jsx/relax.jsx'),
+    slac = require('./jsx/slac.jsx');
 
 // Create new hyphy-vision export
 window.absrel = absrel;
@@ -20,3 +21,4 @@ window.fade = fade;
 window.fade_summary = fade_summary;
 window.relax = relax;
 window.slac = slac;
+window.fel = fel;
