@@ -1,31 +1,28 @@
-var React = require('react'),
-    ReactDOM = require('react-dom');
+var React = require("react");
 
 var InputInfo = React.createClass({
-
-  getInitialState(){
+  getInitialState() {
     return {
       input_data: {}
     };
   },
 
-  componentWillReceiveProps(nextProps){
+  componentWillReceiveProps(nextProps) {
     this.setState({
       input_data: nextProps.input_data
     });
   },
 
-  render(){
-    return(
+  render() {
+    return (
       <div id="input-info">
         <span className="hyphy-highlight">INPUT DATA</span> <span className="divider">/</span>
         <a href="#">{this.state.input_data.filename}</a> <span className="divider">/</span>
         <span className="hyphy-highlight">{this.state.input_data.sequences}</span> sequences <span className="divider">/</span>
         <span className="hyphy-highlight">{this.state.input_data.sites}</span> sites
       </div>
-    )
+    );
   }
-
 });
 
 module.exports.InputInfo = InputInfo;
